requirements.txt
django==2.2
kombu==4.6.3
channels==2.1.7
channels_redis==2.3.3
django-filter==2.2.0
celery==4.3.0
daphne==2.2.0
redis==3.2.1
cryptography==2.7.0
coreapi==2.3.3
djangorestframework==3.9.3
markdown==3.1
django-crispy-forms==1.7.2
django-guardian==1.5.1
django-celery-results==1.1.0
django-crispy-forms==1.7.2
requests==2.19.1
simplejson==3.16.0
django-bootstrap4==0.0.8
flower==0.9.3
mysqlclient==1.4.4
pymysql==0.9.3
django-simple-serializer==2.0.7
django-rest-swagger==2.2.0
django-cors-headers==2.5.1
drf-dynamic-fields==0.3.1
configparser==3.7.4
paramiko==2.6.0
djangorestframework-jwt==1.11.0